refactored and stop using mocks
