+++ src/test/kotlin/no/nav/arbeidsplassen/internalad/indexer/KafkaTopicIndexerTest.kt
@@ -2,81 +2,48 @@ package no.nav.arbeidsplassen.internalad.indexer
 
 import com.github.tomakehurst.wiremock.WireMockServer
 import com.github.tomakehurst.wiremock.client.WireMock.*
-import io.micronaut.configuration.kafka.KafkaProducerFactory
-import io.micronaut.configuration.kafka.annotation.KafkaClient
+import io.micronaut.context.annotation.Primary
 import io.micronaut.context.annotation.Property
-import io.micronaut.context.annotation.PropertySource
 import io.micronaut.context.annotation.Replaces
 import io.micronaut.http.client.DefaultHttpClient
 import io.micronaut.http.client.RxHttpClient
 import io.micronaut.test.annotation.MicronautTest
 import io.micronaut.test.annotation.MockBean
 import io.micronaut.test.support.TestPropertyProvider
-import io.reactivex.Flowable
 import no.nav.arbeidsplassen.internalad.indexer.feed.AdTransport
-import no.nav.arbeidsplassen.internalad.indexer.feed.FeedConnector
-import no.nav.arbeidsplassen.internalad.indexer.index.*
-import org.apache.kafka.clients.consumer.Consumer
+import no.nav.arbeidsplassen.internalad.indexer.index.AdIndexer
+import no.nav.arbeidsplassen.internalad.indexer.index.IndexResponse
+import no.nav.arbeidsplassen.internalad.indexer.index.IndexerService
 import org.apache.kafka.clients.producer.KafkaProducer
 import org.apache.kafka.clients.producer.Producer
 import org.apache.kafka.clients.producer.ProducerRecord
 import org.apache.kafka.common.serialization.StringSerializer
 import org.elasticsearch.rest.RestStatus
 import org.junit.jupiter.api.AfterAll
-import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
-import org.mockito.ArgumentMatchers
-import org.mockito.Mockito.`when`
-import org.mockito.Mockito.mock
-import org.mockito.stubbing.Answer
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.testcontainers.containers.KafkaContainer
-import java.time.LocalDateTime
 import java.util.concurrent.CountDownLatch
 import java.util.concurrent.TimeUnit
-import javax.inject.Inject
+import javax.inject.Singleton
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 @MicronautTest
 class KafkaTopicIndexerTest : TestPropertyProvider {
 
-    @Inject
-    lateinit var adIndexer: AdIndexer
-
     lateinit var kafkaProducer: Producer<String, String>
 
     companion object {
-        private val wireMockServer: WireMockServer = WireMockServer(9001)
         private val LOG: Logger = LoggerFactory.getLogger(KafkaTopicIndexerTest::class.java)
         private val kafkaContainer: KafkaContainer = KafkaContainer()
     }
 
-    val adIndexerLatch = CountDownLatch(1)
+
 
     init {
-        wireMockServer.stubFor(head(urlMatching("/feedtask-internalad.*"))
-                .willReturn(aResponse()
-                        .withStatus(200)))
-        wireMockServer.stubFor(put(urlMatching("/internalad.*"))
-                .willReturn(aResponse()
-                        .withHeader("Content-Type", "application/json")
-                        .withBody("{\"index\": \"foo\", \"acknowledged\": \"true\", \"shards_acknowledged\": \"true\"}")
-                        .withStatus(200)))
-        wireMockServer.stubFor(put(urlMatching("/shedlock.*"))
-                .willReturn(aResponse()
-                        .withHeader("Content-Type", "application/json")
-                        .withBody("{\"index\": \"foo\", \"acknowledged\": \"true\", \"shards_acknowledged\": \"true\"}")
-                        .withStatus(200)))
-        wireMockServer.stubFor(post(urlMatching("/_aliases.*"))
-                .willReturn(aResponse()
-                        .withHeader("Content-Type", "application/json")
-                        .withBody("{\"index\": \"foo\", \"acknowledged\": \"true\", \"shards_acknowledged\": \"true\"}")
-                        .withStatus(200)))
-        wireMockServer.start()
-
         kafkaContainer.start()
         LOG.info("Kafka bootstrap servers: " + kafkaContainer.bootstrapServers)
 
@@ -90,45 +57,36 @@ class KafkaTopicIndexerTest : TestPropertyProvider {
     @BeforeAll
     fun initKafka() {
         val props: Map<String, String> = hashMapOf(Pair("bootstrap.servers", kafkaContainer.bootstrapServers))
-
         kafkaProducer = KafkaProducer<String, String>(props,
             StringSerializer(), StringSerializer())
     }
 
     @AfterAll
-    fun tearDownWiremock() {
-        wireMockServer.shutdownServer()
+    fun tearDown() {
         kafkaContainer.close()
     }
 
     @Test
     @Property(name = "indexer.scheduler.enabled", value = "false")
     fun kafkaListenerTest() {
-        `when`(adIndexer.index(ArgumentMatchers.anyList())).thenAnswer {
-            adIndexerLatch.countDown()
-            IndexResponse(hasFailures = false, status = RestStatus.OK, numItems = 1, failureMessage = "")
-        }
         kafkaProducer.send(ProducerRecord("StillingIntern", "83a08d2c-b0e2-4d5d-88dc-02736b9c6be1", adTransportJSON))
         kafkaProducer.flush()
         val isIndexed = adIndexerLatch.await(20L, TimeUnit.SECONDS)
         assert(isIndexed)
     }
+}
 
-    @MockBean(AdIndexer::class)
-    fun countDownAdIndexer(): AdIndexer {
-        return mock(AdIndexer::class.java)
-    }
-
-    @MockBean(DefaultHttpClient::class)
-    fun mockClient(): RxHttpClient {
-        return mock(RxHttpClient::class.java)
-    }
-
-    @MockBean(IndexerSchedulerTask::class)
-    fun mockSchedulerTask(): IndexerSchedulerTask {
-        return mock(IndexerSchedulerTask::class.java)
+@Primary
+@Replaces(IndexerService::class)
+@Singleton
+class MyIndexerService: AdIndexer {
+    override fun index(ads: List<AdTransport>): IndexResponse {
+        adIndexerLatch.countDown()
+        return IndexResponse(hasFailures = false, status = RestStatus.OK, numItems = 1, failureMessage = "")
     }
 }
+
+val adIndexerLatch = CountDownLatch(1)
 
 val adTransportJSON = """
         {
